feat(boundaries): boundaries implementation
Tackling issue Gab-San/lbm-simulation-lib#32

+++ include/lbm-sim/collision-operators/collision-strategy.hpp
@@ -11,8 +11,8 @@
 
 namespace lbm {
 
-template <int dim, typename VelocitySet, enum CollisionModel cm_t,
-          enum ExecutionBackend backend_t>
+template <unsigned short int dim, typename VelocitySet,
+          enum CollisionModel cm_t, enum ExecutionBackend backend_t>
 class CollisionStrategy {
 public:
   const Params<dim, cm_t> params;
